Updated mail plugin to support skip_verify
Also introduced a mail() VQL function in addition to the plugin
