+++ artifacts/definitions/Server/Alerts/PsExec.yaml
@@ -11,6 +11,9 @@ type: SERVER_EVENT
 parameters:
   - name: EmailAddress
     default: admin@example.com
+  - name: SkipVerify
+    type: bool
+    description: If set we skip TLS verification.
   - name: MessageTemplate
     default: |
       PsExec execution detected at %v: %v for client %v
@@ -28,6 +31,7 @@ sources:
               to=EmailAddress,
               subject='PsExec launched on host',
               period=60,
+              skip_verify=SkipVerify,
               body=format(
               format=MessageTemplate,
               args=[Timestamp, CommandLine, ClientId])
