+++ artifacts/definitions/Server/Monitoring/ClientCount.yaml
@@ -9,6 +9,9 @@ type: SERVER_EVENT
 parameters:
    - name: EmailAddress
      default: admin@example.com
+   - name: SkipVerify
+     type: bool
+     description: If set we skip TLS verification.
    - name: CCAddress
      default:
    - name: Subject
@@ -30,6 +33,7 @@ sources:
                 cc=CCAddress,
                 subject=Subject,
                 period=60,
+                skip_verify=SkipVerify,
                 body=format(format='Total clients currently connected %v',
                      args=[metrics.client_comms_current_connections])
             )
